Messages, hide/show nav, and css
add css for background of app description (transparent dark background)
hide/show for sign-up, sign-in, change-pw, and sign-out so that when a
user is not signed up or signed in or signed out, they are not able to
see the Playlist CRUD action area, and when a user is signed in
successfully, they are able to see the Playlist CRUD action area.
User facing auth messages updated.